% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quaddata.R
\docType{data}
\name{quaddata}
\alias{quaddata}
\title{quad tree data on locations of US blocks (6,246,672 in the Census 2010-based data here)

Classes ‘data.table’ and 'data.frame':	6246672 obs. of  4 variables:
$ BLOCK_X: num  -2442 -2440 -2440 -2439 -2440 ...
$ BLOCK_Z: num  3112 3115 3114 3115 3114 ...
$ BLOCK_Y: num  -144 -144 -144 -144 -143 ...
$ BLOCKID: chr  "0201600010011250" "0201600010011270" "0201600010011291" "0201600010011297" ...}
\description{
quad tree data on locations of US blocks (6,246,672 in the Census 2010-based data here)

Classes ‘data.table’ and 'data.frame':	6246672 obs. of  4 variables:
$ BLOCK_X: num  -2442 -2440 -2440 -2439 -2440 ...
$ BLOCK_Z: num  3112 3115 3114 3115 3114 ...
$ BLOCK_Y: num  -144 -144 -144 -144 -143 ...
$ BLOCKID: chr  "0201600010011250" "0201600010011270" "0201600010011291" "0201600010011297" ...
}
