% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quaddata.R
\docType{data}
\name{quaddata}
\alias{quaddata}
\title{quad tree data on locations of US Census blocks}
\description{
This has selected columns from Census block dataset.
\preformatted{

  Code that created blockdata, quaddata, blockquadtree:
    (also localtree has to be created during runtime. 
    cannot be saved as data file beforehand)
  
  blockdata <- 0
  quaddata <- blockdata[ , .(BLOCK_X, BLOCK_Z, BLOCK_Y, blockid)]
  blockquadtree <- SearchTrees::createTree(quaddata, treeType = "quad", dataType = "point")
  
  2010 data format:
  
  str(quaddata) 
  Classes ‘data.table’ and 'data.frame':	6246672 obs. of  4 variables:
  $ BLOCK_X: num  -2442 -2440 -2440 -2439 -2440 ...
  $ BLOCK_Z: num  3112 3115 3114 3115 3114 ...
  $ BLOCK_Y: num  -144 -144 -144 -144 -143 ...
  $ blockid: chr  "0201600010011250" "0201600010011270" "0201600010011291" "0201600010011297" ...
}
}
\seealso{
blockdata blockquadtree
}
